For a full changelog see https://github.com/Zerotorescue/HoN-Bots. Below are the changes copy-pasted in a random order.
Fixed error with loading HeroData files.
Added CanDispositionSelf, CanDispositionFriendlies and CanDispositionHostiles to AbilityInfo.
Updated HeroData Template to include the new Disposition properties.
Added ShouldAvoidDamage to AbilityInfo.class for abilities like Voodoo Jester's Cursed Ground.
Added HeroData for Voodoo Jester.
Updated HeroData template to make copy-pasting lines easier.
Added SetHeroInfo and GetHeroInfo to AbilityInfo.class so we can get the related HeroInfo when we only know the AbilityInfo. SetHeroInfo is automatically called in HeroInfo:AddAbility(abil).
Added ChannelingState to AbilityInfo. This property is used to specify the state being placed on the hero casting the ability so we can find out what ability is being channeled. This fixes an issue where Magmus' Eruption (R) and Steambath (W) could not be differentiated.
Added GetAllAbilities function to HeroData for retrieving all abilities with the provided filters.
Added IsFrom function to AbilityInfo which can be used to determine if the ability is from the provided unit.
HeroData:GetAllAbilities can now be provided with either a table (for multiple filters), string (for a single filter) or empty (for no filter) filter parameters.